Refactor: invite 를 chat input에서 처리 #120
